Add `yes_no` template macro/helper function
This helps simplify repetitive if/else blocks in the template.

The function by default returns 'Yes' if the passed value is
true or 'No' if it's false.

This is useful in the simple cases like when listing which users are
admin of an S3 bucket, e.g.

    <td>{{ yes_no(users3bucket.is_admin) }}</td>

The function behaviour can be customised to check if the value is equal
to something else, e.g. a string.

This is useful when displaying 'Yes'/'No' depending on the access level
for example:

    <th>Has R/W access?</th>
    <td>{{ yes_no(users3bucket.access_level, 'readwrite') }}</td>

(In the example above the function will compare access level with 'readwrite'
which is considered the "true" value)

The strings returned can be customised so for example you can use the helper
to display something else, e.g.

    <td>{{ yes_no(users3bucket.is_admin, true, '(Admin)', '') }}</td>

Because these strings can be customised, the helper can be used to "toggle"
the access level value, e.g.

    <inpyt type="hidden" name="access_level" value="{{ yes_no(users3bucket.access_level, 'readwrite', 'readonly', 'readwrite') }}" />
    <input type="submit" value="{{ yes_no(users3bucket.access_level, 'readwrite', 'Revoke', 'Grant') }} R/W access" />