Breaking: `url_for()` support for URL fragment
This will allow redirects to a specific tab in the new tabbed homepage.

**NOTE**: This is a breaking change and existing calls will need to change
from `url_for(route, { param_1: value_1 })` to
`url_for(route, { params: { param_1: value_1 } })``.
The spaces are not just to make the linter happy but also to make
jinja happy as `}}` has a special meaning in its context.