Better handling of idled tools
Problem
=======

Tools can now be in an "idled" state (tools are idled overnight to save
resources).

When a tool is idled, its deployment is scaled down so there are no pods
running. This means the CP will not be able to loop through the pops and
show their status.

Another problem is that the "Restart" button is always shown. Users obviously
click on it and expect a tool to wake up from its idled state.

This doesn't work because "Restart" simply delete the pods in a deployment,
causing them to be recreated. When a tool is idled there are zero pods,
so "Restart" is not doing much.

Changes
=======

- I added the `Deployment.idled` property which checks if a tool is idled.
  This returns true if the deployment has the IDLED='true' label
- In the tools page:
  - in the status column, show "Idled" when a tool is idled
  - show the "Unidle" button when a tool is idled. This opens the tool triggering
    the unidling process as usual
  - don't show the "Restart" button when a tool is idled

Ticket
======

https://trello.com/c/am9IhJYr/753-fix-bug-users-cant-recover-from-hibernations-in-the-control-panel

See
===

- Idler: https://github.com/ministryofjustice/analytics-platform-idler/blob/master/idler.py#L126
- Unidler: https://github.com/ministryofjustice/analytics-platform-unidler/blob/master/unidler.py