Improved Tools status polling from page refresh to AJAX request
When the user is on the Tools tab, this panel content is updated every
few seconds with the response from the `GET /tools` endpoint.

This `GET /tools` endpoint renders only the panel content (same template
used by the `home` template) but it's lighter as it only requests the tools
status.