Upgrade github package to octokit/rest
This couldn't happen automatically because of the way we were constructing the
proxy object for the github API.

The old version of the API returned an object that took a config and could be
`extended`, the new version of the api dynamically constructs the rest client by
going through it's own list of plugins and attaching methods to itself. This
didn't play well with the es6 extends syntax and after the version bump the
`authenticate` method we overrode was still pointing to the library version.

I've changed to shadowing it in the constructor instead of overriding it. This
seems to work.