Fix Dependabot PRs
Dependabot generates PRs with branch names containing '/' characters.
We automatically build and tag a Docker image using the branch name as
the tag, but Docker tags may only contain alphanumerics, periods,
dashes and underscores.

This quick fix converts '/' to '-'. This may not always result in a
valid Docker tag, but should cover the vast majority of cases.